refactor, add actions.lua for run actions, evaluate while typing
