+++ lua/nvumi/main.lua
@@ -1,5 +1,5 @@
+local actions = require("nvumi.actions")
 local config = require("nvumi.config")
-local processor = require("nvumi.processor")
 local state = require("nvumi.state")
 local debounce = require("nvumi.debounce").debounce
 local M = {}
@@ -15,8 +15,8 @@ function M.open()
     win_by_ft = {
       nvumi = {
         keys = {
-          ["source"] = { opts.keys.run, processor.run_on_buffer, mode = { "n", "x" }, desc = "Run Numi" },
-          ["reset"] = { opts.keys.reset, processor.reset_buffer, mode = "n", desc = "Reset Buffer" },
+          ["source"] = { opts.keys.run, actions.run_on_buffer, mode = { "n", "x" }, desc = "Run Numi" },
+          ["reset"] = { opts.keys.reset, actions.reset_buffer, mode = "n", desc = "Reset Buffer" },
           ["yank"] = { opts.keys.yank, state.yank_last_output, mode = "n", desc = "Yank Last" },
         },
       },
@@ -44,18 +44,21 @@ function M.setup()
     end,
   })
 
+  -- re-eval when leaving insert mode
+  local debounced_rob = debounce(1000, actions.run_on_buffer)
   vim.api.nvim_create_autocmd({ "InsertLeave", "TextChanged" }, {
     pattern = "*.nvumi",
     callback = function()
-      processor.run_on_buffer()
+      debounced_rob()
     end,
   })
 
-  local debounced_eval = debounce(300, require("nvumi.processor").run_on_line)
+  -- live evaluations while typing
+  local debounced_rol = debounce(500, actions.run_on_line)
   vim.api.nvim_create_autocmd({ "TextChanged", "TextChangedI" }, {
     pattern = "*.nvumi",
     callback = function()
-      debounced_eval()
+      debounced_rol()
     end,
   })
 end
